Module plaquette::sim

Enums

ThermalisationSimumlationError

ThermalizeError

Functions

generate_state_default

Generate a hot configuration with the given config

get_mc_from_config

Generate a [MetropolisHastingsDeltaDiagnostic] from a config

get_mc_from_config_sweep

Generate a [MetropolisHastingsSweep] from a config

get_pb_template

Return the [indicatif::ProgressBar] template

run_simulation_with_progress_bar_average

run_simulation_with_progress_bar_volume

simulation_gather_measurement

thermalize_state

thermalize the state using a observable as a mesurement. It plot a graph and write a csv file for the correlation

thermalize_with_e_field

Type Definitions

ResultThermalizeE